Extracts above the fold components into a common component
